requirements.txt
fastapi
uvicorn[standard]
Pillow
numpy